% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen_binary2.R
\name{datagen_binary2}
\alias{datagen_binary2}
\title{This function generate simulated data for binary outcomes for K=2}
\usage{
datagen_binary2(
  n1,
  n2,
  p,
  s0,
  s1,
  s2,
  rho1,
  rho2,
  sigma1,
  sigma2,
  samesig,
  scale
)
}
\arguments{
\item{n1}{sample size in population 1}

\item{n2}{sample size in population 2}

\item{p}{number of predictors}

\item{s0}{number of true signals}

\item{s1}{number of signals shown in population 1 only}

\item{s2}{number of signals shown in population 2 only}

\item{rho1}{correlation parameter for design matrix in population 1}

\item{rho2}{correlation parameter for design matrix in population 2}

\item{sigma1}{intercept parameter for population 1}

\item{sigma2}{intercept parameter for population 2}

\item{samesig}{whether true signals magnitude the same for all data}
}
\value{
A simulated dataset
}
\description{
This function generate simulated data for binary outcomes for K=2
}
