% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen_con3.R
\name{datagen_con3}
\alias{datagen_con3}
\title{This function generate simulated data for gaussian outcomes for K=3}
\usage{
datagen_con3(
  n1,
  n2,
  n3,
  p,
  s0,
  s1,
  s2,
  s3,
  s12,
  s13,
  s23,
  rho1,
  rho2,
  rho3,
  sigma1,
  sigma2,
  sigma3,
  samesig,
  scale
)
}
\arguments{
\item{n1}{sample size in population 1}

\item{n2}{sample size in population 2}

\item{n3}{sample size in population 3}

\item{p}{number of predictors}

\item{s0}{number of true signals}

\item{s1}{number of signals shown in population 1 only}

\item{s2}{number of signals shown in population 2 only}

\item{s3}{number of signals shown in population 3 only}

\item{s12}{number of signals shown in population 1 and 2 only}

\item{s13}{number of signals shown in population 1 and 3 only}

\item{s23}{number of signals shown in population 2 and 3 only}

\item{rho1}{correlation parameter for design matrix in population 1}

\item{rho2}{correlation parameter for design matrix in population 3}

\item{sigma1}{noise parameter for population 1}

\item{sigma2}{noise parameter for population 2}

\item{sigma3}{noise parameter for population 3}

\item{samesig}{whether true signals magnitude the same for all data}

\item{effect}{scale multiplier}
}
\value{
A simulated dataset
}
\description{
This function generate simulated data for gaussian outcomes for K=3
}
