% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_power_plot.R
\name{FDR_and_power_plot}
\alias{FDR_and_power_plot}
\title{This function creates FDR and power plots}
\usage{
FDR_and_power_plot(
  FDRmat,
  powermat,
  names,
  cols,
  pchs,
  alpha,
  vary_list,
  xlab = "",
  title = "",
  leg_coords = integer(0),
  show_legend = FALSE,
  labelsize = 1,
  legsize = 1,
  cexsize = 1.2
)
}
\description{
This function creates FDR and power plots
}
