% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myest_repfdr.R
\name{myest_repfdr}
\alias{myest_repfdr}
\title{This function uses repfdr method for comparison, only for K=2}
\usage{
myest_repfdr(data, q, family = "gaussian")
}
\arguments{
\item{data}{dataset}

\item{q}{FDR level for control}

\item{family}{type of outcomes from each experiment (gaussian or Cox)}
}
\value{
A list of selected variables
}
\description{
This function uses repfdr method for comparison, only for K=2
}
