% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myest_stack.R
\name{myest_stack}
\alias{myest_stack}
\title{This function implements the Pooling Knockoff method}
\usage{
myest_stack(
  Xlist,
  Ylist,
  familylist,
  Kmethod,
  Zmethod,
  q,
  offset,
  Zfun = NULL,
  IDorder = NULL
)
}
\arguments{
\item{Xlist}{list of length K for predictor matrices from each experiment}

\item{Ylist}{list of length K for outcomes from each experiment}

\item{familylist}{list of length K for type of outcomes from each experiment (Gaussian, Binomial or Cox)}

\item{Kmethod}{list of length K for knockoff construction methods for each experiment (fixed or second order)}

\item{Zmethod}{Current implemented method is glmnet}

\item{q}{FDR level for control}

\item{offset}{0 for Knockoff and 1 for Knockoff+}

\item{Zfun}{User defined function on how to construct Z statistics from a list of length K for Z and a list of length K for Ztilde (required if select Zmethod="Other")}
}
\value{
A list of selected variables
}
\description{
This function implements the Pooling Knockoff method
}
