% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma_AR.R
\name{Sigma_AR}
\alias{Sigma_AR}
\title{This function generates the AR variance covariance matrix with diagonal 1 and correlation rho}
\usage{
Sigma_AR(rho, p)
}
\arguments{
\item{rho}{correlation for adjacent variables}

\item{p}{dimension of the matrix}
}
\value{
A variance-covariance matrix
}
\description{
This function generates the AR variance covariance matrix with diagonal 1 and correlation rho
}
